restructure slightly
