+++ scripts/lint.sh
@@ -1,8 +1,8 @@
 #!/usr/bin/env bash
 
 ## run from top level, i.e. `bash scripts/lint.sh`
-black src/
-black test/
-black commands.py
-black util.py
-black mtriage
+python3 -m black src/
+python3 -m black test/
+python3 -m black commands.py
+python3 -m black util.py
+python3 -m black mtriage
