+++ src/lib/common/util.py
@@ -1,6 +1,23 @@
 import ntpath
 import json
 import hashlib
+import multiprocessing
+
+MAX_CPUS = multiprocessing.cpu_count() - 1
+
+
+def get_batch_size(ls_len):
+    """ Determine the batch size for multiprocessing. """
+    if ls_len >= MAX_CPUS:
+        return ls_len // MAX_CPUS + 1
+    # TODO: improve this heuristic for splitting up jobs
+    return ls_len
+
+
+def batch(iterable, n=1):
+    l = len(iterable)
+    for ndx in range(0, l, n):
+        yield iterable[ndx : min(ndx + n, l)]
 
 
 def serialize_dict(_dict):
